Refactor c++ code
