+++ src/treemap.cpp
@@ -1,92 +1,7 @@
 #include <Rcpp.h>
+#include "nodes.h"
 using namespace Rcpp;
 
-struct Rectangle {
-    double x;
-    double y;
-    double width;
-    double height;
-};
-class Node {
-    std::vector<Node*> children;
-    Node* parent;
-    bool hasParent;
-    double Weight;
-    double Height;
-    int Id;
-    int Order;
-
-    static bool comparePtrToNode(Node* a, Node* b) {
-        return (a->order() < b->order());
-    }
-public:
-    Node() {
-        hasParent = false;
-        Weight = 0.0;
-        Height = 0.0;
-        Id = 0;
-        Order = 0;
-    };
-    Node(int id, int order, double weight) {
-        hasParent = false;
-        Id = id;
-        Order = order;
-        Weight = weight;
-    };
-    Node(int id, int order, double weight, double height) {
-        hasParent = false;
-        Id = id;
-        Order = order;
-        Weight = weight;
-        Height = height;
-    };
-    std::vector<Node*> getChildren() {
-        std::vector<Node*> childVec;
-        for (int i = 0; i < children.size(); ++i) {
-            childVec.push_back(children[i]);
-        }
-        return childVec;
-    };
-    int nChildren() {
-        return children.size();
-    };
-    double weight() {
-        return Weight;
-    };
-    double height() {
-        return Height;
-    };
-    void addNode(Node* n) {
-        double w = n->weight();
-        Weight += w;
-        if (hasParent) {
-            parent->addWeight(w);
-        }
-        n->setParent(this);
-        children.push_back(n);
-    };
-    void addWeight(double w) {
-        if (hasParent) {
-            parent->addWeight(w);
-        }
-        Weight += w;
-    };
-    void setParent(Node* n) {
-        hasParent = true;
-        parent = n;
-    };
-    bool orphan() {
-        return !hasParent;
-    };
-    int order() {
-        return Order;
-    };
-    void sortChildren() {
-        std::sort(children.begin(), children.end(), comparePtrToNode);
-    };
-
-    Rectangle bounds;
-};
 double w(std::vector<Node*>& nodes) {
     double w = 0;
     for (int i = 0; i < nodes.size(); ++i) {
@@ -168,20 +83,6 @@ std::vector<Node*> createHierarchy(std::vector<int> parent, std::vector<int> ord
     }
     return nodes;
 }
-int findTopNode(std::vector<Node*>& nodes) {
-    int i;
-    bool found = false;
-    for (i = 0; i < nodes.size(); ++i) {
-        if (nodes[i]->orphan()) {
-            found = true;
-            break;
-        }
-    }
-    if (!found) {
-        stop("No top node. Is this a tree structure?");
-    }
-    return i;
-}
 
 //[[Rcpp::export]]
 NumericMatrix splitTreemap(IntegerVector parent, IntegerVector order, NumericVector weight, double width, double height) {
@@ -194,15 +95,15 @@ NumericMatrix splitTreemap(IntegerVector parent, IntegerVector order, NumericVec
         nodes[i]->sortChildren();
     }
 
-    int startNode = findTopNode(nodes);
+    Node* startNode = nodes[0]->getRoot();
     Rectangle r = {
         0,
         0,
         width,
         height
     };
-    nodes[startNode]->bounds = r;
-    splitLayout(nodes[startNode]->getChildren(), r);
+    startNode->bounds = r;
+    splitLayout(startNode->getChildren(), r);
 
     for (i = 0; i < nodes.size(); ++i) {
         rect(i, 0) = nodes[i]->bounds.x;
